refactor and more cgroup charts
